requirements.txt
pillow
matplotlib
opencv-contrib-python